+++ addon/components/flexberry-simpleolv.js
@@ -1736,6 +1736,8 @@ export default folv.extend(
         return ['eq', 'neq', 'le', 'ge'];
 
       case 'string':
+        return ['eq', 'neq', 'like'];
+
       case 'boolean':
         return ['eq', 'neq'];
 
@@ -1818,7 +1820,7 @@ export default folv.extend(
         let filters = {};
         let hasFilters = false;
         this.get('columns').forEach((column) => {
-          if (column.filter.pattern && column.filter.condition) {
+          if (column.filter.condition || column.filter.pattern) {
             hasFilters = true;
             filters[column.filter.name] = column.filter;
           }
@@ -1827,7 +1829,11 @@ export default folv.extend(
         if (hasFilters) {
           this.sendAction('applyFilters', filters);
         } else {
-          this.get('currentController').send('refreshList');
+          if (this.get('currentController.filters')) {
+            this.get('currentController').send('resetFilters');
+          } else {
+            this.get('currentController').send('refreshList');
+          }
         }
       } else {
         this.get('currentController').send('refreshList');