+++ blueprints/flexberry-reexport-new/index.js
@@ -1,6 +1,6 @@
 var inflector = require("inflection");
 module.exports = {
-    description: 'Generates an import wrapper.',
+    description: 'Generates an import wrapper for new routes and controllers.',
     availableOptions: [
         { name: 'middle-path', type: String }
     ],
@@ -29,7 +29,9 @@ module.exports = {
     */
     locals: function (options) {
         return {
-            modulePath: options.project.name() + "/" + inflector.pluralize(options.middlePath) + "/" + options.entity.name // for use in files\__root__\__path__\__name__.js
+            middlePath: "" + inflector.pluralize(options.middlePath),
+            projectName: "" + options.project.name(),
+            entityName: "" + options.entity.name
         };
     }
 };